+++ resources/views/application.blade.php
@@ -6,6 +6,7 @@
 
     <!-- CSRF Token -->
     <meta name="csrf-token" content="{{ csrf_token() }}">
+    <meta name="csp-nonce" content="{{ csp_nonce() }}">
     <link type="image/x-icon" rel="icon" href="{{ app(\App\Settings\GeneralSettings::class)->favicon }}">
     <title>{{ app(\App\Settings\GeneralSettings::class)->name }}</title>